Convert floats to doubles

+++ PortlandWeather/DTO/Wind.cs
@@ -5,11 +5,11 @@ namespace PortlandWeather.DTO;
 public class Wind
 {
     [JsonPropertyName("speed")]
-    public required float Speed {  get; set; }
+    public required double Speed {  get; set; }
 
     [JsonPropertyName("deg")]
     public required int Degree { get; set; }
 
     [JsonPropertyName("gust")]
-    public float? Gust { get; set; }
+    public double? Gust { get; set; }
 }
